# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a NodeJS API project to Swagger Documentation",
    "version": "1.0.0",
    "title": "Genome API",
    "contact": {
      "email": "abc@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http","https"],
  "host": "localhost:4000",
  "basePath": "/",
  "securityDefinitions":{
    "bearerAuth":{
      "type":"apiKey",
      "name":"Authorization",
      "in":"header",
      "descriptions":">- Enter the token with the `Bearer:` prefix, e.g. Bearer abcdefg1234" 
    }
      },
  "paths": {
    "/modules/{stationName}": {
      "post": {
        "summary": "Create a new modules if not exist and create new trame",
        "description": "Create a new modules if not exist and create new trame",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
            {
                "name": "stationName",
                "in": "path",
                "description": "module name that needs to be create or to create the trame",
                "required": true,
                "type": "string"
            },
          {
            "in": "body",
            "name": "body",
            "description": "trame object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "temp": {
                  "type": "string"
                },
                "fuel": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "ph1": {
                  "type": "string"
                },
                "ph2": {
                  "type": "string"
                },
                "ph3": {
                  "type": "string"
                },
                "freq": {
                  "type": "string"
                },
                "oilPress": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/moduleCreatedResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/modules/{idModule}": {
        "get": {
          "security":[{
            "bearerAuth":[]
          }],
          "summary": "Obtenir les informations sur un le groupe donc l'identifiant est idModule",
          "description": "Obtenir les informations sur un le groupe donc l'identifiant est idModule",
          "consumes": ["application/json"],
          "parameters": [
              {
                  "name": "idModule",
                  "in": "path",
                  "description": "Id module that needs to get informations",
                  "required": true,
                  "type": "string"
              }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/modulesResponse"
                }
              }
            },
            "500": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            },
            "401": {
              "description": "Unauthorized!",
              "schema": {
                "$ref": "#/definitions/UnautorizedError"
              }
            },
            "403": {
              "description": "Role Unauthorized!",
              "schema": {
                "$ref": "#/definitions/PermissionError"
              }
            }
          }
        }
      },
      "/modules/search": {
        "get": {
          "summary": "Obtenir les informations sur un le groupe donc l'identifiant est idModule",
          "description": "Obtenir les informations sur un le groupe donc l'identifiant est idModule",
          "consumes": ["application/json"],
          "parameters": [
              {
                  "name": "name",
                  "in": "query",
                  "description": "Nom du groupe ou stationName",
                  "required": true,
                  "type": "string"
              },
              {
                "name": "page",
                "in": "query",
                "description": "Numero de la page",
                "required": false,
                "type": "integer"
            },
            {
                "name": "limit",
                "in": "query",
                "description": "Nombre maximun d'element",
                "required": false,
                "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/modulesResponse"
                }
              }
            },
            "500": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      },
      "/modules/{idModule}/trames": {
        "get": {
          "summary": "Obtenir toutes les trames d'un module",
          "description": "Obtenir toutes les trames d'un module",
          "consumes": ["application/json"],
          "parameters": [
              {
                  "name": "idModule",
                  "in": "path",
                  "description": "id module or generator that needs to get all trames",
                  "required": true,
                  "type": "string"
              },
              {
                "name": "page",
                "in": "query",
                "description": "Numero de la page",
                "required": false,
                "type": "integer"
            },
            {
                "name": "limit",
                "in": "query",
                "description": "Nombre d'element à afficher dans la page",
                "required": false,
                "type": "integer"
            },
            {
                "name": "from",
                "in": "query",
                "description": "Date à partir de laquelle on veut l'information",
                "required": false,
                "type": "string"
            },
            {
                "name": "at",
                "in": "query",
                "description": "Date de fin à laquelle on veut l'information",
                "required": false,
                "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/TramesResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      },
      "/modules/{idModule}/vidanges": {
        "get": {
          "summary": "Obtenir toutes les vidanges d'un module",
          "description": "Obtenir toutes les vidanges d'un module",
          "consumes": ["application/json"],
          "parameters": [
              {
                  "name": "idModule",
                  "in": "path",
                  "description": "id module or generator that needs to get all trames",
                  "required": true,
                  "type": "string"
              },
              {
                "name": "page",
                "in": "query",
                "description": "Numero de la page",
                "required": false,
                "type": "integer"
            },
            {
                "name": "limit",
                "in": "query",
                "description": "Nombre d'element à afficher dans la page",
                "required": false,
                "type": "integer"
            },
            {
                "name": "from",
                "in": "query",
                "description": "Date à partir de laquelle on veut l'information",
                "required": false,
                "type": "string"
            },
            {
                "name": "at",
                "in": "query",
                "description": "Date de fin à laquelle on veut l'information",
                "required": false,
                "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/vidangesResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      },
      "/modules": {
        "get": {
          "summary": "Obtenir tous les modules",
          "description": "Obtenir tous les modules",
          "consumes": ["application/json"],
          "security":[{
            "bearerAuth":[]
          }],
          "parameters": [
              {
                "name": "page",
                "in": "query",
                "description": "Numero de la page",
                "required": false,
                "type": "integer"
            },
            {
                "name": "limit",
                "in": "query",
                "description": "Nombre d'element à afficher dans la page",
                "required": false,
                "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/modulesResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            },
            "401": {
              "description": "Unauthorized!",
              "schema": {
                "$ref": "#/definitions/UnautorizedError"
              }
            },
            "403": {
              "description": "Role Unauthorized!",
              "schema": {
                "$ref": "#/definitions/PermissionError"
              }
            }
          }
        }
      },
...
  },
  "definitions": {
...
    "modulesResponse":{
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description":"Identifiant du module"
          },
          "stationName": {
            "type": "string",
            "description":"Nom de la station"
          },
          "position": {
            "type": "object",
            "description":"information dur la position geographique lat et long",
            "properties":{
            "lat":{
                "type":"string",
                "description":"latitude"
            },
            "long":{
                "type":"string",
                "description":"longitude"
            }
            }
          },
          "dataRequired":{
            "type":"object",
            "properties":{
            "temp":{
                "type":"string",
                "description":"Temperature minimun"
            },
            "fuel":{
                "type":"string"
            },
            "ph1":{
                "type":"string"
            },
            "ph2":{
                "type":"string"
            },
            "ph3":{
                "type":"string"
            },
            "oilPress":{
                "type":"string"
            },
            "freq":{
                "type":"string"
            }
            }
          },
          "elapse":{
            "type":"object",
            "description":"Durée de fonctionnement depuis la derniere vidange",
            "properties":{
            "milliseconds":{
                "type":"string",
                "description":"Millisecondes"
            },
            "seconds":{
                "type":"string"
            },
            "minutes":{
                "type":"string"
            },
            "hours":{
                "type":"string"
            },
            "ph3":{
                "type":"string"
            },
            "days":{
                "type":"string"
            },
            "months":{
                "type":"string"
            },
            "years":{
                "type":"string"
            }
            }
          },
          "elapse_total":{
            "type":"object",
            "description":"Durée de fonctionnement depuis le prémier démarrage du groupe",
            "properties":{
            "milliseconds":{
                "type":"string",
                "description":"Millisecondes"
            },
            "seconds":{
                "type":"string"
            },
            "minutes":{
                "type":"string"
            },
            "hours":{
                "type":"string"
            },
            "ph3":{
                "type":"string"
            },
            "days":{
                "type":"string"
            },
            "months":{
                "type":"string"
            },
            "years":{
                "type":"string"
            }
            }
          },
          "infoVidange":{
            "type":"object",
            "description":"Durée de fonctionnement depuis le prémier démarrage du groupe",
            "properties":{
            "id":{
                "type":"string",
                "description":"Identifiant de la vidange"
            },
            "remarque":{
                "type":"string",
                "description":"Remarque faites lors de la vidange"
            },
            "date":{
                "type":"string",
                "description":"Date de la vidange"
            },
            "createdAt":{
                "type":"string",
                "description":"Date de creation coté serveur"
            },
            "ph3":{
                "type":"string"
            },
            "days":{
                "type":"string"
            },
            "idModule":{
                "type":"string",
                "description":"Identifiant du groupe qui a été vidangé"
            }
            }
          },
          "elapse_total_hours_minutes":{
            "type":"object",
            "properties":{
              "hours":{
                "type":"integer",
                "description":"Nombre d'heure de fonctionnement"
              },
              "minutes":{
                "type":"integer",
                "description":"Minutes"
              }
            },
            "description":"Durée de fonctionnement du groupe format heures et minutes"
          },
          "elapse_hours_minutes":{
            "type":"object",
            "properties":{
              "hours":{
                "type":"integer",
                "description":"Nombre d'heure de fonctionnement"
              },
              "minutes":{
                "type":"integer",
                "description":"Minutes"
              }
            },
            "description":"Durée de fonctionnement du groupe depuis la derniére vidange format heures et minutes"
          }
        }
    },
...
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
      "moduleCreatedResponse":{
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "idModule": {
            "type": "string",
            "description":"Identifiant du module"
          }
        }
      },
...
      "vidangesResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description":"Identifiant de la vidange"
          },
          "remarque": {
            "type": "string",
            "description":"Remarques lors de la vidange"
          },
          "date": {
            "type": "string",
            "description":"Date de la vidange"
          },
          "createdAt": {
            "type": "string",
            "description":"Date de la creation de la vidange coté serveur"
          },
          "idModule": {
            "type": "string",
            "description":"Identifiant du module ou du generateur qui a été vidangé"
          },
          "doBy": {
            "type": "string",
            "description":"L'identifiant de celui qui a fait la vidange"
          }
        }
      },
...
      "UnautorizedError":{
        "description":"Access token is missing or invalid"
      },
      "PermissionError":{
        "description":"User's role is not authorized"
      }
  }
}